Fix expenses ability to destroy an object
When a user, allowed to delete by permissions page, want to delete an Expense, piwam denied it.

It seems that controller action was not correctly mapped : delete use destroy.

After this commit, when the delete permission is checked, user is allowed to destroy Expense.

Note: IFAICS, it seems other objects are affected too.